# batocera-settings: Re-implement in C