--- a/package/batocera/core/batocera-scripts/src/.clang-format
+++ b/package/batocera/core/batocera-scripts/src/.clang-format
@@ -1,0 +1,1 @@
+BasedOnStyle: Google