--- a/package/batocera/core/batocera-scripts/src/batocera_read_file.h
+++ b/package/batocera/core/batocera-scripts/src/batocera_read_file.h
@@ -1,0 +1,10 @@
+
+#ifndef _BATOCERA_READ_FILE_H_
+#define _BATOCERA_READ_FILE_H_
+
+#include <stddef.h>
+
+/* Reads the file. Exits on error. */
+void batocera_read_file(const char *fname, char **out_data, size_t *out_size);
+
+#endif  // _BATOCERA_READ_FILE_H_