--- a/package/batocera/core/batocera-scripts/src/batocera_read_file.h
+++ b/package/batocera/core/batocera-scripts/src/batocera_read_file.h
@@ -1,4 +1,3 @@
-
 #ifndef _BATOCERA_READ_FILE_H_
 #define _BATOCERA_READ_FILE_H_