… package/batocera/core/batocera-scripts/src/batocera_stringbuf.h
… package/batocera/core/batocera-scripts/src/meson.build
… package/batocera/core/batocera-scripts/src/test/batocera-settings-test.sh
… package/batocera/core/batocera-scripts/src/test/batocera-test.conf