Re-implements `batocera-settings` in C, as two separate binaries for getting and setting the values.

```
Usage: batocera-settings-get [-f CONFIG_FILE] <KEY> [KEY]...

Prints the value of the key or returns a non-zero exit status.
If multiple keys are given, tries them in order until it finds a key that exists.

By default, reads from /userdata/system/batocera.conf
```

```
Usage: batocera-settings-set [-f CONFIG_FILE] <KEY> <VALUE> [KEY VALUE]...

Sets value(s) in the config file.
If a commented key exists, will uncomment it and set the value in-place instead of appending a new key

By default, writes to /userdata/system/batocera.conf
```

Note: This is safe to merge because this PR does not migrate any callers. That will be done in a separate PR.