@glebm 2020-12-05
@lbrpdx Done

Benchmark: Reading a key from a small config file 100 times on my machine: 0.41s -> 0.06s

* Before: 0.4s

  ```bash
  time { for i in $(seq 1 100); do ../../scripts/batocera-settings -f ../test/batocera-test.conf -r kodi.xbutton > /dev/null; done }
  ```

* After: 0.06s

   ```bash
   time { for i in $(seq 1 100); do ./batocera-settings-get -f ../test/batocera-test.conf kodi.xbutton > /dev/null; done }
   ```